# expression: implement vectorized evaluation for `builtinUnaryNot*` series of functions