<!--
Thank you for contributing to TiDB! Please read TiDB's [CONTRIBUTING](https://github.com/pingcap/community/blob/master/CONTRIBUTING.md) document **BEFORE** filing this PR.
-->

### What problem does this PR solve? <!--add issue link with summary if exists-->

Implement vectorized evaluation for the series of unary not functions:

- `builtinUnaryNotRealSig`
- `builtinUnaryNotDecimalSig`
- `builtinUnaryNotIntSig`

Issue: #12103 

### What is changed and how it works?

For `builtinUnaryNotRealSig`, about 8 times faster:

```
BenchmarkVectorizedBuiltinOpFunc/builtinUnaryNotRealSig-VecBuiltinFunc-4    	  500000	      3111 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinOpFunc/builtinUnaryNotRealSig-NonVecBuiltinFunc-4 	  100000	     24126 ns/op	       0 B/op	       0 allocs/op
```

For `builtinUnaryNotDecimalSig`, about 3 times faster:

```
BenchmarkVectorizedBuiltinOpFunc/builtinUnaryNotDecimalSig-VecBuiltinFunc-4 	  200000	      8971 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinOpFunc/builtinUnaryNotDecimalSig-NonVecBuiltinFunc-4         	   50000	     27541 ns/op	       0 B/op	       0 allocs/op
```

For `builtinUnaryNotIntSig`, about 8 times faster:

```
BenchmarkVectorizedBuiltinOpFunc/builtinUnaryNotIntSig-VecBuiltinFunc-4                	  500000	      2996 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinOpFunc/builtinUnaryNotIntSig-NonVecBuiltinFunc-4             	   50000	     25370 ns/op	       0 B/op	       0 allocs/op
```


### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

 - Unit test