@k-ye 2019-09-28
Would be happy to split this into three PRs, if that is more compliant with the contribution rules!