@qw4990 2019-09-29
> Would be happy to split this into three PRs, if that is more compliant with the contribution rules!

I think it's ok since these three functions are relatively simple.